+++ server/src/scimodom/services/user.py
@@ -2,7 +2,6 @@ import logging
 import random
 import string
 from functools import cache
-from typing import Optional
 
 from sqlalchemy import select
 from sqlalchemy.orm import Session
@@ -17,14 +16,20 @@ logger = logging.getLogger(__name__)
 
 
 class UserExists(Exception):
+    """Exception for handling existing users."""
+
     pass
 
 
 class WrongUserOrPassword(Exception):
+    """Exception for handling failed login."""
+
     pass
 
 
 class NoSuchUser(Exception):
+    """Exception for handling non-existing user."""
+
     pass
 
 
@@ -33,11 +38,12 @@ class _DetailedWrongUserOrPassword(Exception):
 
 
 class UserService:
-    """Service to handle users. Beside checking a password on
-    login it supports the workflows to register a user and
-    to rest a password. All errors, which may caused by hacker
-    attacks are logged in detailed but reported to the outside
-    as a WrongUserOrPassword exception with a generic message.
+    """Service to handle users.
+
+    Provide password checking on login, password reset and
+    registration workflow. Possible hacking alerts are logged,
+    but reported to the user as a WrongUserOrPassword exception
+    with a generic message.
 
     :param session: SQLAlchemy ORM session
     :type session: Session
@@ -53,15 +59,16 @@ class UserService:
         self._mail_service = mail_service
 
     def register_user(self, email: str, password: str) -> None:
-        """Create a new, inactive user in the database and send out
-        a token to validate the email address. It may fail with a
-        UserExists exception.
+        """Register a user.
 
-        :param email: A user is identified by the email address. There is no
-        separate name.
+        Create a new, inactive user in the database and send out
+        a token to validate the email address.
+
+        :param email: User email address. This is the user name.
         :type email: str
         :param password: Clear text password
         :type password: str
+        :raises UserExists: If user already exists.
         """
         try:
             self.get_user_by_email(email)
@@ -87,11 +94,12 @@ class UserService:
             raise exc
 
     def get_user_by_email(self, email: str) -> User:
-        """Get user.
+        """Get user by email.
 
         :param email: User name (email address).
         :type email: str
-        :returns: User instance
+        :raises NoSuchUser: If user does not exist.
+        :return: User instance
         :rtype: User
         """
         stmt = select(User).where(User.email == email)
@@ -104,20 +112,17 @@ class UserService:
             raise NoSuchUser(f"No such user '{email}'")
         return users[0]
 
-    def _get_random_token(self):
-        return "".join(random.choice(self.TOKEN_CHARACTERS) for _ in range(32))
+    def confirm_user(self, email: str, confirmation_token: str) -> None:
+        """Activate a registered user using the email token.
 
-    def _send_email_with_confirmation_token(self, email, confirmation_token):
-        pass
-
-    def confirm_user(self, email: str, confirmation_token: str):
-        """Activates a registered user with the token sent out before by email.
-        If the user is active already just ignore it and count it as success.
+        If the user is already active, nothing is done.
 
         :param email: User name (email address).
         :type email: str
         :param confirmation_token: Token
         :type confirmation_token: str
+        :raises WrongUserOrPassword: Failed attempt, covers up
+        a _DetailedWrongUserOrPassword exception.
         """
         try:
             try:
@@ -143,11 +148,12 @@ class UserService:
             raise WrongUserOrPassword("Wrong username or password!")
 
     def request_password_reset(self, email: str) -> None:
-        """A token is generated and send out by email. Otherwise, the account stays
-        unchanged, e.g. inactive or active. That is important, because otherwise an
-        unauthenticated hacker may abuse the workflow to trigger a state change of
-        the account. The workflow can also be used to retry registration if the
-        initial email with the token was lost.
+        """Generate a token and send out by email.
+
+        The account stays unchanged, e.g. inactive or active. This is important
+        because otherwise an unauthenticated hacker may abuse the workflow
+        to trigger a state change of the account. The workflow can also be
+        used to retry registration if the initial email with the token is lost.
 
         :param email: User name (email address).
         :type email: str
@@ -168,6 +174,8 @@ class UserService:
         :type confirmation_token: str
         :param new_password: New password
         :type new_password: str
+        :raises WrongUserOrPassword: Failed attempt, covers up
+        a _DetailedWrongUserOrPassword exception.
         """
         try:
             try:
@@ -186,13 +194,14 @@ class UserService:
             logger.warning(f"{str(exc)}")
             raise WrongUserOrPassword("Wrong username or password!")
 
-    def change_password(self, email: str, new_password: str):
+    def change_password(self, email: str, new_password: str) -> None:
         """Change a password for an authorized user.
 
         :param email: User name (email address).
         :type email: str
         :param new_password: New password
         :type new_password: str
+        :raises _DetailedWrongUserOrPassword: If user does not exists.
         """
         try:
             user = self.get_user_by_email(email)
@@ -206,8 +215,11 @@ class UserService:
         self._session.commit()
 
     def check_password(self, email: str, password: str) -> bool:
-        """Returns true if the password matches the stored password.
-        Otherwise, false is returned - also in case of an unknown or inactive user.
+        """Check password.
+
+        Returns true if the password matches the stored password.
+        Otherwise, false is returned - also in the case of an unknown or
+        inactive user.
 
         :param email: User name (email address).
         :type email: str
@@ -229,12 +241,15 @@ class UserService:
             return False
         return True
 
+    def _get_random_token(self) -> str:
+        return "".join(random.choice(self.TOKEN_CHARACTERS) for _ in range(32))
+
 
 @cache
-def get_user_service():
-    """Helper function to set up a UserService object by injecting its dependencies.
+def get_user_service() -> UserService:
+    """Instantiate a UserService object by injecting its dependencies.
 
-    :returns: User service instance
+    :return: User service instance
     :rtype: UserService
     """
     return UserService(session=get_session(), mail_service=get_mail_service())